handling coverageStatements in PkgUpdateTIPPs
moved updateCoverage() to TippService
