+++ server/gokbg3/grails-app/services/org/gokb/TippService.groovy
@@ -6,6 +6,9 @@ import com.k_int.ConcurrencyManagerService.Job
 import grails.converters.JSON
 import org.gokb.cred.Combo
 import org.gokb.cred.IdentifierNamespace
+import org.gokb.cred.RefdataValue
+import org.gokb.cred.TIPPCoverageStatement
+import org.gokb.rest.TippController
 import org.grails.web.json.JSONObject
 import org.gokb.cred.KBComponent
 import org.gokb.cred.RefdataCategory
@@ -14,6 +17,7 @@ import org.gokb.cred.TitleInstancePackagePlatform
 import org.gokb.cred.Package
 
 import java.time.LocalDateTime
+import java.time.ZoneId
 
 
 class TippService {
@@ -22,6 +26,98 @@ class TippService {
   def sessionFactory
   def reviewRequestService
   def autoTimestampEventListener
+
+  public def updateCoverage(tipp, reqBody) {
+    def cov_list = reqBody.coverageStatements ?: reqBody.coverage
+    def missing = tipp.coverageStatements.collect { it.id }
+    def changed = false
+
+    cov_list?.each { c ->
+      def parsedStart = GOKbTextUtils.completeDateString(c.startDate)
+      def parsedEnd = GOKbTextUtils.completeDateString(c.endDate, false)
+
+      changed |= com.k_int.ClassUtils.setStringIfDifferent(tipp, 'startVolume', c.startVolume)
+      changed |= com.k_int.ClassUtils.setStringIfDifferent(tipp, 'startIssue', c.startIssue)
+      changed |= com.k_int.ClassUtils.setStringIfDifferent(tipp, 'endVolume', c.endVolume)
+      changed |= com.k_int.ClassUtils.setStringIfDifferent(tipp, 'endIssue', c.endIssue)
+      changed |= com.k_int.ClassUtils.setStringIfDifferent(tipp, 'embargo', c.embargo)
+      changed |= com.k_int.ClassUtils.setStringIfDifferent(tipp, 'coverageNote', c.coverageNote)
+      changed |= com.k_int.ClassUtils.setDateIfPresent(parsedStart, tipp, 'startDate')
+      changed |= com.k_int.ClassUtils.setDateIfPresent(parsedEnd, tipp, 'endDate')
+      changed |= com.k_int.ClassUtils.setRefdataIfPresent(c.coverageDepth, tipp, 'coverageDepth', 'TitleInstancePackagePlatform.CoverageDepth')
+
+      def cs_match = false
+      def startAsDate = (parsedStart ? Date.from(parsedStart.atZone(ZoneId.systemDefault()).toInstant()) : null)
+      def endAsDate = (parsedEnd ? Date.from(parsedEnd.atZone(ZoneId.systemDefault()).toInstant()) : null)
+
+      tipp.coverageStatements?.each { tcs ->
+
+        if (!cs_match && (
+          (c.id && tcs.id == c.id) ||
+            (tcs.startVolume && tcs.startVolume == c.startVolume) ||
+            (tcs.startDate && tcs.startDate == startAsDate) ||
+            (!cs_match && !tcs.startVolume && !tcs.startDate && !tcs.endVolume && !tcs.endDate))
+        ) {
+          changed |= com.k_int.ClassUtils.setStringIfDifferent(tcs, 'startIssue', c.startIssue)
+          changed |= com.k_int.ClassUtils.setStringIfDifferent(tcs, 'startVolume', c.startVolume)
+          changed |= com.k_int.ClassUtils.setStringIfDifferent(tcs, 'endVolume', c.endVolume)
+          changed |= com.k_int.ClassUtils.setStringIfDifferent(tcs, 'endIssue', c.endIssue)
+          changed |= com.k_int.ClassUtils.setStringIfDifferent(tcs, 'embargo', c.embargo)
+          changed |= com.k_int.ClassUtils.setStringIfDifferent(tcs, 'coverageNote', c.coverageNote)
+          changed |= com.k_int.ClassUtils.setDateIfPresent(parsedStart, tcs, 'startDate')
+          changed |= com.k_int.ClassUtils.setDateIfPresent(parsedEnd, tcs, 'endDate')
+
+          cs_match = true
+          missing.remove(tcs.id)
+        }
+        else if (cs_match) {
+          TippController.log.debug("Matched new coverage ${c} on multiple existing coverages!")
+        }
+      }
+
+      if (!cs_match) {
+
+        def cov_depth = null
+
+        if (c.coverageDepth instanceof String) {
+          cov_depth = RefdataCategory.lookup('TIPPCoverageStatement.CoverageDepth', c.coverageDepth)
+        }
+        else if (c.coverageDepth instanceof Integer) {
+          cov_depth = RefdataValue.get(c.coverageDepth)
+        }
+        else if (c.coverageDepth instanceof Map) {
+          if (c.coverageDepth.id) {
+            cov_depth = RefdataValue.get(c.coverageDepth.id)
+          }
+          else {
+            cov_depth = RefdataCategory.lookup('TIPPCoverageStatement.CoverageDepth', (c.coverageDepth.name ?: c.coverageDepth.value))
+          }
+        }
+
+        if (!cov_depth) {
+          cov_depth = RefdataCategory.lookup('TIPPCoverageStatement.CoverageDepth', "Fulltext")
+        }
+
+        tipp.addToCoverageStatements('startVolume': c.startVolume,  \
+           'startIssue': c.startIssue,  \
+           'endVolume': c.endVolume,  \
+           'endIssue': c.endIssue,  \
+           'embargo': c.embargo,  \
+           'coverageDepth': cov_depth,  \
+           'coverageNote': c.coverageNote,  \
+           'startDate': startAsDate,  \
+           'endDate': endAsDate
+        )
+      }
+    }
+    if (cov_list) {
+      missing.each {
+        tipp.removeFromCoverageStatements(TIPPCoverageStatement.get(it))
+      }
+    }
+
+    tipp
+  }
 
   def matchPackage(Package aPackage) {
